remove get in favor of blob get
